29.05.2020 12:02

+++ AutoRepair/ViewModel/ClientTabViewModel.cs
@@ -1,18 +1,30 @@
 ﻿using System;
-using System.Collections.ObjectModel;
 using System.Linq;
+using System.Reactive;
+using System.Reactive.Linq;
+using System.Windows;
 using AutoRepair.Model;
+using AutoRepair.View;
 using DynamicData.Binding;
 using Microsoft.EntityFrameworkCore;
 using ReactiveUI;
-using System.Reactive;
-using System.Reactive.Linq;
-using AutoRepair.View;
 
 namespace AutoRepair.ViewModel
 {
     internal class ClientTabViewModel : ReactiveObject
     {
+        #region IsClientSelectedProperty
+
+        private IObservable<bool> IsClientSelected => this.WhenAnyValue(x => x.SelectedClient).Select(x => x != null);
+
+        #endregion
+
+        #region ClientsProperty
+
+        public ObservableCollectionExtended<Client> Clients { get; set; }
+
+        #endregion
+
         #region Constructors
 
         public ClientTabViewModel()
@@ -21,28 +33,17 @@ namespace AutoRepair.ViewModel
             EditClientCommand = ReactiveCommand.Create(EditClient, IsClientSelected);
             DeleteClientCommand = ReactiveCommand.Create(DeleteClient, IsClientSelected);
             UpdateDatabaseEvent.DatabaseUpdated += DataBaseUpdated;
-            Clients=new ObservableCollectionExtended<Client>();
+            Clients = new ObservableCollectionExtended<Client>();
             DataBaseUpdated();
         }
 
         private void DataBaseUpdated()
         {
-            using (AppContext db=new AppContext())
+            using (AppContext db = new AppContext())
             {
-                Clients.Load(db.Clients.Include(x=>x.ClientCars).ThenInclude(x=>x.CarModel));
+                Clients.Load(db.Clients.Include(x => x.ClientCars).ThenInclude(x => x.CarModel));
             }
         }
-
-
-        #endregion
-
-        #region IsClientSelectedProperty
-        private IObservable<bool> IsClientSelected => this.WhenAnyValue(x => x.SelectedClient).Select(x => x != null);
-        #endregion
-
-        #region ClientsProperty
-
-        public ObservableCollectionExtended<Client> Clients { get; set; }
 
         #endregion
 
@@ -59,38 +60,55 @@ namespace AutoRepair.ViewModel
         #endregion
 
         #region AddClientCommand
+
         public ReactiveCommand<Unit, Unit> AddClientCommand { get; }
+
         private void AddClient()
         {
             new ClientEditWindow().ShowDialogAsync();
-            MessageBus.Current.SendMessage(0,"WindowMode");
+            MessageBus.Current.SendMessage(0, "WindowMode");
         }
+
         #endregion
 
         #region EditClientCommand
+
         public ReactiveCommand<Unit, Unit> EditClientCommand { get; }
+
         private void EditClient()
         {
             new ClientEditWindow().ShowDialogAsync();
-            MessageBus.Current.SendMessage(1,"WindowMode");
-            MessageBus.Current.SendMessage(SelectedClient,"EditClient");
+            MessageBus.Current.SendMessage(1, "WindowMode");
+            MessageBus.Current.SendMessage(SelectedClient, "EditClient");
         }
+
         #endregion
 
         #region DeleteClientCommand
+
         public ReactiveCommand<Unit, Unit> DeleteClientCommand { get; }
+
         private void DeleteClient()
         {
-            using (AppContext db=new AppContext())
+            MessageBoxResult messageBoxResult = MessageBox.Show(
+                "Вы действительно хотите удалить клиента" + Environment.NewLine +
+                "ВНИМАНИЕ!! Удаление клиента привод к удалению всех его машин и заказов.", "Удалить?",
+                MessageBoxButton.YesNo);
+            if (messageBoxResult == MessageBoxResult.Yes)
             {
-                Client client = db.Clients.Include(x=>x.ClientCars).First(x => x.ClientID == SelectedClient.ClientID);
+                using (AppContext db = new AppContext())
+                {
+                    Client client = db.Clients.Include(x => x.ClientCars)
+                        .First(x => x.ClientID == SelectedClient.ClientID);
 
-                db.Clients.Remove(client);
-                db.SaveChanges();
+                    db.Clients.Remove(client);
+                    db.SaveChanges();
+                }
+
+                UpdateDatabaseEvent.OnDatabaseUpdated();
             }
-            DataBaseUpdated();
         }
-        #endregion
 
+        #endregion
     }
 }
